move helper function location and fix test config
