+++ test/sagemaker_tests/tensorflow/tensorflow2_training/integration/sagemaker/test_experiments.py
@@ -26,7 +26,7 @@ from smexperiments.trial_component import TrialComponent
 from ...integration import DEFAULT_TIMEOUT
 from ...integration import RESOURCE_PATH
 from .timeout import timeout
-from .... import invoke_tensorflow_estimator
+from ... import invoke_tensorflow_estimator
 
 DATA_PATH = os.path.join(RESOURCE_PATH, "mnist")
 SCRIPT_PATH = os.path.join(DATA_PATH, "mnist_gluon_basic_hook_demo.py")
