+++ test/sagemaker_tests/tensorflow/tensorflow2_training/integration/sagemaker/test_horovod.py
@@ -19,7 +19,7 @@ import pytest
 import sagemaker
 
 from ...integration.utils import processor, py_version, unique_name_from_base  # noqa: F401
-from .... import invoke_tensorflow_estimator
+from ... import invoke_tensorflow_estimator
 
 RESOURCE_PATH = os.path.join(os.path.dirname(__file__), '..', '..', 'resources')
