+++ test/sagemaker_tests/tensorflow/tensorflow2_training/integration/sagemaker/test_tuning_model_dir.py
@@ -18,7 +18,7 @@ import pytest
 
 from sagemaker.tensorflow import TensorFlow
 from sagemaker.tuner import HyperparameterTuner, IntegerParameter
-from .... import invoke_tensorflow_estimator
+from ... import invoke_tensorflow_estimator
 
 from ...integration.utils import processor, py_version, unique_name_from_base  # noqa: F401
